// app/interfaces/piezas.ts
export interface Piezas {
    action: any;
    reporte?: any;
    cantPieza1?: any;
    numPieza1?: any;
    nombrePieza1?: any;
    cantPieza2?: any;
    numPieza2?: any;
    nombrePieza2?: any;
    cantPieza3?: any;
    numPieza3?: any;
    nombrePieza3?: any;
    cantPieza4?: any;
    numPieza4?: any;
    nombrePieza4?: any;
    cantPieza5?: any;
    numPieza5?: any;
    nombrePieza5?: any;
}